Fixed filenames & directory structure, deleted unused imports
